DiffuQwen/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.9.0
deepspeed>=0.13.0
qwen-vl-utils>=0.0.2
pillow>=10.0.0
pdf2image>=1.16.0
pyyaml>=6.0
tqdm>=4.66.0
einops>=0.7.0
safetensors>=0.4.0
